DEVPROD-8712 Remove legacy spawn host routes
